style_requirement.txt
kenlm
fasttext
configargparse
torchtext==0.7.0
streamtologger